laboratorio_farmacia/requirements.txt
scikit-learn==1.3.0 
pandas==2.0.3 
pyyaml==6.0 
joblib==1.3.0 
